@@ -1,6 +1,6 @@
 {
     "name": "makeit-search",
-    "version": "1.1.1",
+    "version": "1.1.2",
     "author": {
         "url": "https://www.makeit.vip",
         "name": "lirongtong",
@@ -18,7 +18,7 @@
         "makeit search",
         "麦可易特网"
     ],
-    "description": "Makeit Search 搜索组件，是基于 Vue3.0 + Vite 开发，默认根据本地数据进行数据筛选，支持远程搜索，同时也支持搜索结果分页配置，自定义结果列表中每一项的点击事件等",
+    "description": "Makeit Search 搜索组件，是基于 Vue3.x + Vite3.x 开发，默认根据本地数据进行数据筛选，支持远程搜索，同时也支持搜索结果分页配置，自定义结果列表中每一项的点击事件等",
     "private": false,
     "license": "MIT",
     "repository": {
@@ -47,9 +47,10 @@
         "lint": "eslint -c ./.eslintrc --fix --ext .jsx,.js,.vue,.ts ./src"
     },
     "dependencies": {
-        "@ant-design/icons-vue": "^5.1.9",
-        "axios": "^0.21.1",
-        "vue": "^3.0.5"
+        "@ant-design/icons-vue": "^6.1.0",
+        "axios": "^0.27.2",
+        "vue": "^3.2.37",
+        "vue-types": "^4.1.1"
     },
     "devDependencies": {
         "@babel/core": "^7.12.10",
@@ -64,21 +65,26 @@
         "@babel/plugin-transform-typescript": "^7.12.1",
         "@babel/preset-env": "^7.12.11",
         "@babel/preset-typescript": "^7.12.7",
-        "@typescript-eslint/eslint-plugin": "^4.13.0",
-        "@typescript-eslint/parser": "^4.13.0",
+        "@typescript-eslint/eslint-plugin": "^5.27.0",
+        "@typescript-eslint/parser": "^5.27.0",
+        "@vitejs/plugin-vue": "^2.3.3",
+        "@vitejs/plugin-vue-jsx": "^1.3.10",
         "@vue/babel-plugin-jsx": "^1.0.1",
-        "@vue/compiler-sfc": "^3.0.5",
+        "@vue/compiler-sfc": "^3.2.37",
         "@vue/eslint-config-prettier": "^6.0.0",
-        "@vue/eslint-config-typescript": "^7.0.0",
+        "@vue/eslint-config-typescript": "^10.0.0",
         "autoprefixer": "^9.8.6",
         "babel-loader": "^8.2.2",
         "case-sensitive-paths-webpack-plugin": "^2.3.0",
         "clean-webpack-plugin": "^3.0.0",
         "compression-webpack-plugin": "^7.1.2",
         "css-loader": "^5.0.1",
-        "eslint": "^7.17.0",
-        "eslint-plugin-prettier": "^3.3.1",
-        "eslint-plugin-vue": "^7.4.1",
+        "css-minimizer-webpack-plugin": "^4.0.0",
+        "eslint": "^8.11.0",
+        "eslint-config-prettier": "^8.5.0",
+        "eslint-plugin-import": "^2.26.0",
+        "eslint-plugin-prettier": "^3.4.1",
+        "eslint-plugin-vue": "^8.5.0",
         "gulp": "^4.0.2",
         "gulp-autoprefixer": "^7.0.1",
         "gulp-babel": "^8.0.0",
@@ -93,7 +99,6 @@
         "less-plugin-npm-import": "^2.1.0",
         "merge2": "^1.4.1",
         "mini-css-extract-plugin": "^1.3.4",
-        "optimize-css-assets-webpack-plugin": "^5.0.4",
         "postcss": "^8.2.4",
         "postcss-loader": "^4.1.0",
         "prettier": "^2.2.1",
@@ -101,15 +106,15 @@
         "terser-webpack-plugin": "^5.1.1",
         "through2": "^4.0.2",
         "ts-loader": "^8.0.14",
-        "typescript": "^4.1.3",
+        "typescript": "^4.7.4",
         "url-loader": "^4.1.1",
-        "vite": "^1.0.0-rc.13",
+        "vite": "^2.9.14",
+        "vite-plugin-eslint": "^1.6.1",
         "vue-loader": "^16.1.2",
         "vue-style-loader": "^4.1.2",
         "vue-template-compiler": "^2.6.12",
-        "vue-types": "^3.0.1",
-        "webpack": "^5.14.0",
-        "webpack-cli": "^4.3.1",
-        "webpack-merge": "^5.7.3"
+        "webpack": "^5.73.0",
+        "webpack-cli": "^4.9.2",
+        "webpack-merge": "^5.8.0"
     }
 }
