@@ -1,7 +1,7 @@
 import { App, Plugin } from 'vue'
 import Search from './search'
 
-Search.install = function(app: App) {
+Search.install = function (app: App) {
     app.component(Search.name, Search)
     app.component(Search.Key.name, Search.Key)
     return app
